backend/requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2
numpy==1.24.0
accelerate==0.24.0
pandas==2.1.0